refactor: move the tree rendering and the popovers to app-render.js (#33)
renderRepositories, renderPullRequests, renderPullRequest, renderParticipant,
renderOrphanedIssues and their helpers move, unchanged, into a module that
knows nothing about the filter state. initializePopovers follows them: it
reads the data attributes renderPullRequest writes. app.js imports what it
uses; nothing else changes. The HTML rendered from the fixture data of both
projects is identical before and after.
